#6 refactoring pr반영
