+++ routes/serviceCenter.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const router = express.Router();
 const serviceCenterCtrl = require("../controllers/serviceCenter");
-const multerUpload = require("../utils/multer");
 
 router.post("/notice/create", serviceCenterCtrl.noticeCreate);
 router.get("/notice/list", serviceCenterCtrl.noticeList);
